refactor: data is now better centralized
