--- src/components/Hobbies/MusicData.ts
@@ -1,19 +1,0 @@
-export interface Genre {
-  title: string;
-  body: string;
-}
-
-export const genres: Genre[] = [
-  {
-    title: "J-Pop",
-    body: "Due to my japanese language studies, I've been listening to a lot of japanese songs, which is one of the genres that I enjoy the most at the moment.",
-  },
-  {
-    title: "Rock",
-    body: "What would I be without rock? I still remember the first time that I listened to 'Duality' from Slipknot, it got me instantly hooked, I'm not gonna lie.",
-  },
-  {
-    title: "Alternative",
-    body: "I'm calling it alternative just because I believe some of the songs that I listen are not too mainstream, but it could really be anything that I vibe with.",
-  },
-];
