+++ src/pages/AboutMe.tsx
@@ -13,7 +13,7 @@ import {
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
-    root: {
+    card: {
       display: "flex",
     },
     details: {
@@ -25,6 +25,8 @@ const useStyles = makeStyles((theme: Theme) =>
     },
     media: {
       width: 150,
+      height: 150,
+      alignSelf: "center",
     },
     strong: {
       color: theme.palette.primary.main,
@@ -51,7 +53,7 @@ export default function AboutMe() {
               <Container>
                 <Grid container spacing={2}>
                   <Grid item md={6}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -82,7 +84,7 @@ export default function AboutMe() {
                     </Card>
                   </Grid>
                   <Grid item md={6}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -127,7 +129,7 @@ export default function AboutMe() {
               <Container>
                 <Grid container spacing={2}>
                   <Grid item md={6}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -156,7 +158,7 @@ export default function AboutMe() {
                     </Card>
                   </Grid>
                   <Grid item md={6}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -199,7 +201,7 @@ export default function AboutMe() {
               <Container>
                 <Grid container spacing={2}>
                   <Grid item md={4}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -215,7 +217,7 @@ export default function AboutMe() {
                     </Card>
                   </Grid>
                   <Grid item md={4}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -231,7 +233,7 @@ export default function AboutMe() {
                     </Card>
                   </Grid>
                   <Grid item md={4}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -265,7 +267,7 @@ export default function AboutMe() {
               <Container>
                 <Grid container spacing={2}>
                   <Grid item md={4}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -281,7 +283,7 @@ export default function AboutMe() {
                     </Card>
                   </Grid>
                   <Grid item md={4}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
@@ -297,7 +299,7 @@ export default function AboutMe() {
                     </Card>
                   </Grid>
                   <Grid item md={4}>
-                    <Card className={classes.root} elevation={0}>
+                    <Card className={classes.card} elevation={0}>
                       <div className={classes.details}>
                         <CardContent className={classes.content}>
                           <Typography variant="h5" gutterBottom>
